fix: mobile friendly content metadata
